create auth page and movied  some files
