+++ helper/useLocale.ts
@@ -1,8 +1,8 @@
 import { useRouter } from "next/router";
-import translations from "../translation/text.json";
+// import translations from "../translation/text.json";
 
-export const useLocale = () => {
+export const useLocale = (props:any) => {
 	const { locale } = useRouter();
-  const t = locale === "en" ? translations.en.setting : translations.jp.setting;
+  const t = locale === "en" ? props.en : props.jp;
 	return { locale, t };
 };
