+++ pages/_app.tsx
@@ -1,15 +1,15 @@
 import type { AppProps } from "next/app";
-import AuthContext from "../Context/AuthContext";
 import TotalContext from "Context/TotalContext";
 import { SessionProvider } from "next-auth/react";
+import Layout from "@components/Layout";
 export default function App({ Component, pageProps }: AppProps) {
-  return (
-    <SessionProvider session={pageProps.session}>
-      <AuthContext>
-        <TotalContext>
-          <Component {...pageProps} />
-        </TotalContext>
-      </AuthContext>
-    </SessionProvider>
-  );
+	return (
+		<SessionProvider session={pageProps.session}>
+			<Layout>
+				<TotalContext>
+					<Component {...pageProps} />
+				</TotalContext>
+			</Layout>
+		</SessionProvider>
+	);
 }
