+++ pages/api/jwt.ts
@@ -8,6 +8,6 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
 	// if using `NEXTAUTH_SECRET` env variable, we detect it, and you won't actually need to `secret`
 	// const token = await getToken({ req })
 	const token = await getToken({ req, secret });
-	console.log("JSON Web Token", token);
+	// console.log("JSON Web Token", token);
 	res.end()
 }
