+++ pages/home/text.json
@@ -1,0 +1,6 @@
+{
+	"en": {
+		"headline": "Create your own invoice so that you’ll never miss your income",
+    "getstarted":"Get Started"
+	}
+}
