+++ pages/login/text.json
@@ -1,0 +1,7 @@
+{
+  "en":{
+    "headline":"Welcom back",
+    "desc": "Gald to see you again!!",
+    "btn":"Please Sign-in with your Google here."
+  }
+}
